fix cloud connection, mod schemaless use one connection
