+++ taos-ws/Cargo.toml
@@ -21,6 +21,7 @@ futures = { version = "0.3" }
 itertools = "0.13.0"
 log = "0.4"
 once_cell = "1"
+parse_duration = "2.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1" }
 serde_repr = "0.1.8"
@@ -42,9 +43,8 @@ flume = "0.11"
 cfg-if = "1"
 
 [dependencies.tokio-tungstenite]
-version = "0.23"
 git = "https://github.com/taosdata/tokio-tungstenite.git"
-rev = "19822e0"
+rev = "3420353"
 
 [dev-dependencies]
 pretty_env_logger = "0.5.0"
@@ -63,3 +63,6 @@ native-tls-vendored = ["tokio-tungstenite/native-tls-vendored", "native-tls"]
 sync = []
 tmq = []
 deflate = ["tokio-tungstenite/deflate", "tokio-tungstenite/rustls-dangerous"]
+
+rustls-aws-lc-crypto-provider = ["rustls", "tokio-tungstenite/rustls-aws-lc-crypto-provider"]
+rustls-ring-crypto-provider = ["rustls", "tokio-tungstenite/rustls-ring-crypto-provider"]
