+++ taos-ws/src/query/asyn.rs
@@ -324,6 +324,14 @@ async fn read_queries(
                 let ws2 = ws2.clone();
                 let (req_id, data, ok) = v.ok();
                 match &data {
+                    WsRecvData::Insert(_) => {
+                        if let Some((_, sender)) = queries_sender.remove(&req_id) {
+                            sender.send(ok.map(|_| data)).unwrap();
+                        } else {
+                            debug_assert!(!queries_sender.contains(&req_id));
+                            log::warn!("req_id {req_id} not detected, message might be lost");
+                        }
+                    }
                     WsRecvData::Query(_) => {
                         if let Some((_, sender)) = queries_sender.remove(&req_id) {
                             if let Err(err) = sender.send(ok.map(|_| data)) {
@@ -1148,6 +1156,28 @@ impl WsTaos {
         }
     }
 
+    pub async fn s_put(&self, sml: &SmlData) -> RawResult<()> {
+        let action = WsSend::Insert {
+            protocol: sml.protocol() as u8,
+            precision: sml.precision().into(),
+            data: sml.data().join("\n").to_string(),
+            ttl: sml.ttl(),
+            req_id: sml.req_id(),
+        };
+        log::trace!("put send: {:?}", action);
+        let req = self.sender.send_recv(action).await?;
+
+        match req {
+            WsRecvData::Insert(res) => {
+                log::trace!("put resp : {:?}", res);
+                Ok(())
+            }
+            _ => {
+                unreachable!()
+            }
+        }
+    }
+
     pub fn version(&self) -> &str {
         &self.sender.version.version
     }
@@ -1310,8 +1340,8 @@ impl AsyncQueryable for WsTaos {
             .await
     }
 
-    async fn put(&self, _data: &SmlData) -> RawResult<()> {
-        todo!()
+    async fn put(&self, data: &SmlData) -> RawResult<()> {
+        self.s_put(data).in_current_span().await
     }
 }
